Token 생성 시 시간 값 누락 오류 수정

+++ src/main/kotlin/com/retro/retromall/member/dto/TokenAttributes.kt
@@ -3,6 +3,6 @@ package com.retro.retromall.member.dto
 data class TokenAttributes(
     var grantType: String? = null,
     var accessToken: String? = null,
-    var expiredAccessToken: Int? = null,
-    var expiredRefreshToken: Int? = null
+    var expiredAccessToken: Long? = null,
+    var expiredRefreshToken: Long? = null
 )
